requirements.txt
python-dotenv
pydantic
langchain-core
fastapi
uvicorn
pandas
SQLAlchemy
openai
langgraph==0.0.40
langchain
langchain-openai
